Enforcer fixes.
