+++ src/Enforcers/Test/InsertModelValidationEnforcerTest.php
@@ -1,33 +1,13 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+namespace PDGA\DataObjects\Enforcers\Test;
 
 use PDGA\DataObjects\Attributes\Column;
 use PDGA\DataObjects\Enforcers\InsertModelValidationEnforcer;
 use PDGA\DataObjects\Validators\EmailValidator;
 use PDGA\DataObjects\Validators\MaxLengthValidator;
 use PDGA\Exception\ValidationListException;
-
-class InsertedPerson
-{
-    #[Column(
-        name: 'email',
-        maxLength: 30,
-    )]
-    #[MaxLengthValidator(15), EmailValidator]
-    public ?string $email;
-
-    #[Column(
-        name: 'id',
-        isGenerated: true,
-    )]
-    public int $id;
-
-    #[Column(
-        name: 'name',
-    )]
-    public string $name;
-}
+use PHPUnit\Framework\TestCase;
 
 class InsertModelValidationEnforcerTest extends TestCase
 {
