+++ src/Enforcers/Test/ModelValidationEnforcerTest.php
@@ -1,30 +1,12 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+namespace PDGA\DataObjects\Enforcers\Test;
 
 use PDGA\DataObjects\Attributes\Column;
 use PDGA\DataObjects\Enforcers\ModelValidationEnforcer;
 use PDGA\DataObjects\Validators\EmailValidator;
 use PDGA\DataObjects\Validators\MaxLengthValidator;
-
-class NamedPerson
-{
-    #[Column(
-        name: 'email',
-        maxLength: 30
-    )]
-    #[MaxLengthValidator(15), EmailValidator]
-    public ?string $email;
-
-    public int $id;
-
-    #[Column(
-        name: 'name',
-        maxLength: 20
-    )]
-    #[MaxLengthValidator(15)]
-    public string $name;
-}
+use PHPUnit\Framework\TestCase;
 
 class ModelValidationEnforcerTest extends TestCase
 {
