+++ src/Enforcers/Test/MutateModelValidationEnforcerTest.php
@@ -1,34 +1,13 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+namespace PDGA\DataObjects\Enforcers\Test;
 
 use PDGA\DataObjects\Attributes\Column;
 use PDGA\DataObjects\Enforcers\MutateModelValidationEnforcer;
 use PDGA\DataObjects\Validators\EmailValidator;
 use PDGA\DataObjects\Validators\MaxLengthValidator;
 use PDGA\Exception\ValidationListException;
-
-class MutatedPerson
-{
-    #[Column(
-        name: 'email',
-        maxLength: 30,
-    )]
-    #[MaxLengthValidator(15), EmailValidator]
-    public ?string $email;
-
-    #[Column(
-        name: 'id',
-        isGenerated: true,
-        isPrimary: true,
-    )]
-    public int $id;
-
-    #[Column(
-        name: 'name',
-    )]
-    public string $name;
-}
+use PHPUnit\Framework\TestCase;
 
 class MutateModelValidationEnforcerTest extends TestCase
 {
